[cute] Move promotion function to cute/device_ir_lowering
stack-info: PR: https://github.com/pytorch/helion/pull/2613, branch: hinriksnaer/stack/14
