+++ helion/_compiler/cute/device_ir_lowering.py
@@ -8,13 +8,11 @@ from typing import cast
 import torch
 
 from ..compile_environment import CompileEnvironment
-from ..device_ir import collect_cute_half_atomic_output_promotions
-from ..device_ir import promote_cute_root_graph_host_tensors
-from ..device_ir import rewrite_cute_half_atomic_output_allocations
 from ..device_ir_lowering import DeviceIRLowering
 
 if TYPE_CHECKING:
     from ..device_ir import DeviceIR
+    from ..device_ir import GraphInfo
     from ..host_function import HostFunction
 
 
@@ -50,6 +48,161 @@ def _promote_cute_half_atomic_outputs(device_ir: DeviceIR) -> None:
         host_fn = HostFunction.current()
         rewrite_cute_half_atomic_output_allocations(host_fn, promotions)
         promote_cute_root_graph_host_tensors(device_ir.graphs, promotions)
+
+
+def collect_cute_half_atomic_output_promotions(
+    graph_infos: list[GraphInfo],
+) -> dict[str, torch.dtype]:
+    from ...language import atomic_add
+    from ...language._tracing_ops import _host_tensor
+    from ..host_function import HostFunction
+    from ..variable_origin import ArgumentOrigin
+
+    promotions: dict[str, torch.dtype] = {}
+    host_fn = HostFunction.current()
+    host_tensor_nodes: dict[str, list[torch.fx.Node]] = {}
+
+    for graph_info in graph_infos:
+        for node in graph_info.graph.nodes:
+            if node.op == "call_function" and node.target is _host_tensor:
+                target_name = node.args[0]
+                if isinstance(target_name, str):
+                    host_tensor_nodes.setdefault(target_name, []).append(node)
+
+    def is_promotable_target(node: torch.fx.Node) -> bool:
+        target_val = node.meta.get("val")
+        if (
+            not isinstance(target_val, torch.Tensor)
+            or target_val.dtype != torch.float16
+        ):
+            return False
+        origin = host_fn.tensor_to_origin.get(target_val)
+        if origin is None or isinstance(origin, ArgumentOrigin):
+            return False
+        if not node.users:
+            return False
+        for user in node.users:
+            if user.op != "call_function" or user.target is not atomic_add:
+                return False
+            if len(user.args) < 3 or user.args[0] is not node or len(user.users) != 0:
+                return False
+            value_node = user.args[2]
+            if not isinstance(value_node, torch.fx.Node):
+                return False
+            value_val = value_node.meta.get("val")
+            if not isinstance(value_val, torch.Tensor) or value_val.dtype not in (
+                torch.float16,
+                torch.float32,
+            ):
+                return False
+        return True
+
+    for target_name, nodes in host_tensor_nodes.items():
+        if all(is_promotable_target(node) for node in nodes):
+            promotions[target_name] = torch.float16
+
+    return promotions
+
+
+def rewrite_cute_half_atomic_output_allocations(
+    host_fn: HostFunction,
+    promotions: dict[str, torch.dtype],
+) -> None:
+    import ast
+
+    from ..ast_extension import create
+    from ..ast_extension import expr_from_string
+
+    torch_factory_names = {
+        "empty",
+        "empty_like",
+        "full",
+        "full_like",
+        "ones",
+        "ones_like",
+        "zeros",
+        "zeros_like",
+    }
+
+    def dtype_expr(dtype: torch.dtype) -> ast.expr:
+        expr = expr_from_string(f"torch.{str(dtype).split('.', 1)[1]}")
+        assert isinstance(expr, ast.expr)
+        return expr
+
+    def is_torch_factory_call(call: ast.Call) -> bool:
+        return (
+            isinstance(call.func, ast.Attribute)
+            and call.func.attr in torch_factory_names
+            and isinstance(call.func.value, ast.Name)
+            and call.func.value.id == "torch"
+        )
+
+    def rewrite_allocation_dtype(call: ast.Call) -> None:
+        dtype = dtype_expr(torch.float32)
+        for kwarg in call.keywords:
+            if kwarg.arg == "dtype":
+                kwarg.value = dtype
+                return
+        if is_torch_factory_call(call):
+            call.keywords.append(create(ast.keyword, arg="dtype", value=dtype))
+
+    def rewrite_return_expr(expr: ast.expr) -> ast.expr:
+        if isinstance(expr, ast.Name) and expr.id in promotions:
+            cast_expr = expr_from_string(
+                "{value}.to({dtype})",
+                value=expr,
+                dtype=dtype_expr(promotions[expr.id]),
+            )
+            assert isinstance(cast_expr, ast.expr)
+            return cast_expr
+        if isinstance(expr, ast.Tuple):
+            return create(
+                ast.Tuple,
+                elts=[rewrite_return_expr(elt) for elt in expr.elts],
+                ctx=expr.ctx,
+            )
+        if isinstance(expr, ast.List):
+            return create(
+                ast.List,
+                elts=[rewrite_return_expr(elt) for elt in expr.elts],
+                ctx=expr.ctx,
+            )
+        return expr
+
+    for stmt in ast.walk(ast.Module(body=host_fn.body, type_ignores=[])):
+        if (
+            isinstance(stmt, ast.Assign)
+            and len(stmt.targets) == 1
+            and isinstance(stmt.targets[0], ast.Name)
+            and stmt.targets[0].id in promotions
+            and isinstance(stmt.value, ast.Call)
+        ):
+            rewrite_allocation_dtype(stmt.value)
+        elif isinstance(stmt, ast.Return) and stmt.value is not None:
+            stmt.value = rewrite_return_expr(stmt.value)
+
+
+def promote_cute_root_graph_host_tensors(
+    graph_infos: list[GraphInfo],
+    promotions: dict[str, torch.dtype],
+) -> None:
+    from ...language._tracing_ops import _host_tensor
+    from ..host_function import HostFunction
+
+    host_fn = HostFunction.current()
+    for graph_info in graph_infos:
+        for node in graph_info.graph.nodes:
+            if node.op != "call_function" or node.target is not _host_tensor:
+                continue
+            target_name = node.args[0]
+            if not isinstance(target_name, str) or target_name not in promotions:
+                continue
+            value = node.meta.get("val")
+            if isinstance(value, torch.Tensor):
+                promoted_value = value.to(dtype=torch.float32)
+                if origin := host_fn.tensor_to_origin.get(value):
+                    host_fn.tensor_to_origin[promoted_value] = origin
+                node.meta["val"] = promoted_value
 
 
 def _register_cute_vector_widths(device_ir: DeviceIR) -> None:
